SNOW-187770 Support to write large partition with less memory for AWS
SNOW-187770
If the data size is bigger than the chunk size, AWS multiple parts upload
API will be used to upload with less memory.
If the data size is smaller than it, current normal upload API is still used.
NOTE: this change only affects writing to AWS with internal stage.