SNOW-202025: Add DateFunctionsSuite to IT Tests (#300)
This PR adds DateFunctionsSuite to the connector's IT tests. The suite is more-or-less a direct copy of Spark's DateFunctionsSuite, with the Seq.toDF() implicit conversion overridden to create a Snowflake Dataframe.

There are 8 failures in the test discovered by porting these tests over. These are tracked in:
https://snowflakecomputing.atlassian.net/browse/SNOW-205533

For now, these tests are disabled in SFDateFunctionsSuite.


Co-authored-by: Edward Ma <etduwx@gmail.com>